Fix for broken Ofsted page
Since we've updated how we display who created a version (as part of the
work to keep deleted user notes and versions), we needed to update Ofsted item
as they also use PaperTrail versions. The fix is to use the same Version
model that the ServiceVersion inherits from, so that the user
association is available for Ofsted items.